agent/requirements.txt
requests
PyYAML
pyfiglet


